evaluation/metrics/evaluate_pronunciation.py
import parselmouth
import librosa
from pathlib import Path
from typing import Dict, Any

def evaluate_pronunciation(sentence_data: Dict[str, Any], audio_file_path: Path) -> Dict[str, Any]:
    """Evaluates the pronunciation accuracy of synthesized audio.

    This function aims to compare the phonetic realization in the audio against
    an expected pronunciation, potentially derived from the input text using
    tools like phonemizers or pronunciation dictionaries.

    Args:
        sentence_data: Dictionary containing sentence information (text, id, etc.).
        audio_file_path: Path to the generated audio file (.wav).

    Returns:
        A dictionary containing pronunciation evaluation results.
        Example: {'phonetic_accuracy': 0.9, 'mispronounced_phonemes': [...]}
        (Placeholder: Returning dummy data for now)
    """
    print(f"Evaluating pronunciation for: {sentence_data.get('id')} ({audio_file_path.name})")

    if not audio_file_path.exists():
        print(f"Error: Audio file not found at {audio_file_path}")
        return {'error': 'Audio file not found'}

    try:
        # Load audio (optional, might only need path for external tools)
        y, sr = librosa.load(audio_file_path, sr=None)
        print(f"Loaded audio: {len(y)} samples, Sample rate: {sr} Hz")

        # --- Placeholder Pronunciation Analysis ---
        # TODO: Implement actual pronunciation analysis. This could involve:
        # 1. Generating reference phonemes for sentence_data['text'] using a phonemizer.
        # 2. Performing forced alignment using Parselmouth/Praat or another tool
        #    to get recognized phonemes and timings from the audio.
        # 3. Comparing reference and recognized phonemes (e.g., Phoneme Error Rate).

        # Example dummy result
        result = {
            'phonetic_accuracy_score': 0.85, # Dummy value
            'details': "Placeholder evaluation - requires phonetic analysis implementation."
        }
        # -----------------------------------------

    except Exception as e:
        print(f"Error during pronunciation evaluation for {audio_file_path.name}: {e}")
        return {'error': str(e)}

    return result

if __name__ == '__main__':
    # Example Usage (requires a dummy audio file)
    print("Running pronunciation evaluation example...")
    dummy_sentence = {
        'id': 'sent_dummy',
        'text': 'This is a dummy sentence for testing.'
    }
    # Create a dummy wav file for testing if it doesn't exist
    from scipy.io.wavfile import write
    import numpy as np
    dummy_audio_path = Path("evaluation/dummy_audio.wav")
    if not dummy_audio_path.exists():
        print("Creating dummy audio file...")
        dummy_audio_path.parent.mkdir(parents=True, exist_ok=True)
        samplerate = 22050
        duration = 2 # seconds
        frequency = 440 # Hz
        t = np.linspace(0., duration, int(samplerate * duration))
        amplitude = np.iinfo(np.int16).max * 0.5
        data = amplitude * np.sin(2. * np.pi * frequency * t)
        write(dummy_audio_path, samplerate, data.astype(np.int16))
        print(f"Dummy audio file created at {dummy_audio_path}")
    else:
        print(f"Using existing dummy audio file at {dummy_audio_path}")

    results = evaluate_pronunciation(dummy_sentence, dummy_audio_path)
    print(f"Pronunciation Results: {results}")
