requirements.txt
jiwer
librosa
praat-parselmouth 